add search to filters
